pyproject.toml
[tool.poetry]
name = "airbnb_clone_back"
version = "0.1.0"
description = ""
authors = ["bh224 <ericakim224@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.2"
django-environ = "^0.9.0"
Pillow = "^9.3.0"
djangorestframework = "^3.14.0"
PyJWT = "^2.6.0"
django-cors-headers = "^3.13.0"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]
black = {version = "^22.10.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
